Rename missing reaction_role occurances to message_to_role
